Server/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-core==2.20.1
